refactor: change folder name to plural
