+++ src/components/common/CodeSheet.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@/components/ui/button";
 import { Sheet, SheetContent, SheetTrigger } from "@/components/ui/sheet";
-import { highlightCode } from "@/lib/highlightCode";
+import { highlightCode } from "@/libs/highlightCode";
 import { type ReactNode, useEffect, useState } from "react";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "../ui/tabs";
 import "@/styles/shikiStyles.css";
