+++ src/components/common/DialogNav.tsx
@@ -1,5 +1,5 @@
 import { ScrollArea, ScrollBar } from "@/components/ui/scroll-area";
-import { cn } from "@/lib/utils";
+import { cn } from "@/libs/utils";
 import { useEffect } from "react";
 import { Link, useLocation } from "react-router-dom";
