+++ src/components/ui/form.tsx
@@ -11,7 +11,7 @@ import {
 } from "react-hook-form";
 
 import { Label } from "@/components/ui/label";
-import { cn } from "@/lib/utils";
+import { cn } from "@/libs/utils";
 
 const Form = FormProvider;
