+++ src/components/ui/inline-dialog.tsx
@@ -1,5 +1,5 @@
 import { Button, buttonVariants } from "@/components/ui/button";
-import { cn } from "@/lib/utils";
+import { cn } from "@/libs/utils";
 import { XIcon } from "lucide-react";
 import type React from "react";
 import {
