+++ src/libs/helmetwrapper.tsx
@@ -1,4 +1,4 @@
-import { siteConfig } from "@/config/site";
+import { siteConfig } from "@/statics/site";
 import { Helmet, HelmetProvider } from "react-helmet-async";
 
 interface HelmetWrapperProps {
